refactor: creating utils folder
