+++ channel-program/src/utils/formatDate.js
@@ -1,0 +1,9 @@
+// Date and time formats can be localized
+export const DATE_FORMAT = "L"; // Ex: 12/13/2019
+export const TIME_FORMAT = "LT"; // Ex: 8:17 AM
+export const DATE_TIME_FORMAT = "LT L"; // Ex: 8:17 AM 12/13/2019
+export const TIME_DATE_FORMAT = "L LT"; // Ex: 12/13/2019 8:17 AM
+export const DATE_TIME_FULL_FORMAT = "LLLL"; // Ex: Friday, December 13, 2019 8:17 AM
+export function formatDate(dateToFormat, format) {
+    return ;
+}
